requirements.txt
streamlit
PyPDF2
pytesseract
Pillow
nltk